+++ db2/core2/signer_rule.go
@@ -9,7 +9,7 @@ type SignerRule struct {
 	ID        uint64                 `db:"id"`
 	Resource  xdr.SignerRuleResource `db:"resource"`
 	Action    string                 `db:"action"`
-	IsForbid  bool                   `db:"is_forbid"`
+	Forbids   bool                   `db:"forbids"`
 	IsDefault bool                   `db:"is_default"`
 	OwnerID   string                 `db:"owner_id"`
 	Details   regources.Details      `db:"details"`